api/requirements.txt
absl-py==0.13.0
alabaster==0.7.12
aniso8601==8.1.0
APScheduler==3.7.0
asgiref==3.2.7
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.4.1
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.6.20
chardet==3.0.4
clang==5.0
click==7.1.2
clifford==1.3.1
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
doccano-transformer==1.0.2
docutils==0.16
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
gunicorn==20.0.4
h5py==3.1.0
haversine==2.3.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imbalanced-learn==0.7.0
importlib-metadata==3.3.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
llvmlite==0.34.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mpl-toolkits.clifford==0.0.3
murmurhash==1.0.2
mysql-connector==2.2.9
networkx==2.5
nltk==3.6.2
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.5
pandas==1.1.4
parso==0.7.1
pathy==0.6.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
plotly==4.6.0
plyer==2.0.0
preshed==3.0.2
prompt-toolkit==3.0.7
protobuf==3.18.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.7.4
Pygments==2.7.0
pylint==2.6.0
pyodbc==4.0.30
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
PyWavelets==1.1.1
pywin32==228
pyzmq==19.0.2
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==1.0.1
scipy==1.5.3
seaborn==0.11.1
six==1.15.0
sklearn==0.0
smart-open==5.1.0
snowballstemmer==2.0.0
soupsieve==2.0.1
spacy==3.0.6
spacy-legacy==3.0.6
spacytextblob==3.0.1
sparse==0.11.2
Sphinx==1.7.1
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
sqlparse==0.3.1
srsly==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
textblob==0.15.3
thinc==8.0.6
threadpoolctl==2.1.0
tifffile==2020.10.1
Tinkerer==1.7.2
toml==0.10.2
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.4
trimesh==3.8.14
typed-ast==1.4.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
Unipath==1.1
urllib3==1.25.10
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.0
